pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "act0r"
version = "0.1.0"
description = "Local-first security testing tool for LLM agents"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.7,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
